+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## [master](https://github.com/anmonteiro/lumo/compare/1.5.0...HEAD) (unreleased)
 
+
+### New features
+
+- Add ability to execute a script from standard input ([#168](https://github.com/anmonteiro/lumo/pull/168)).
+
 ### Changes
 
 - Upgrade ClojureScript to version 1.9.562.